refactor: create `IconButton` component and add hover state styles
