+++ src/app/(main)/my-voyage/[teamId]/sprints/page.tsx
@@ -1,6 +1,6 @@
 // import VoyageSubmittedMessage from "./components/VoyageSubmittedMessage";
 // import EmptyState from "./components/EmptyState";
-import Agenda from "./components/Agenda";
+import Agenda from "./components/agenda/Agenda";
 import MeetingOverview from "./components/MeetingOverview";
 import SprintActions from "./components/SprintActions";
 import Banner from "@/components/banner/Banner";
